requirements.txt
# Core LLM / Agent Framework
crewai
langchain
langchain-community
langchain-core
langchain-google-genai

# Vector Database
chromadb

# Embeddings / LLM Providers (adjust based on what you're using)
openai
tiktoken

# Environment Management
python-dotenv

# API Layer (if serving via API)
fastapi
uvicorn

# Validation & Schemas
pydantic

# YAML Config Loading
pyyaml

# Data Processing
pandas

# Optional: Async + HTTP (often required by LLM libs)
httpx

# Integrations
atlassian-python-api  # Confluence connector
pysnow                # ServiceNow connector (optional)

# Crew ai core package
crewai